+++ src/neuroconv/datainterfaces/__init__.py
@@ -83,6 +83,8 @@ from .ophys.caiman.caimandatainterface import CaimanSegmentationInterface
 from .ophys.cnmfe.cnmfedatainterface import CnmfeSegmentationInterface
 from .ophys.extract.extractdatainterface import ExtractSegmentationInterface
 from .ophys.hdf5.hdf5datainterface import Hdf5ImagingInterface
+from .ophys.inscopix.inscopixsegmentationdatainterface import InscopixSegmentationInterface
+from .ophys.inscopix.inscopiximagingdatainterface import InscopixImagingInterface
 from .ophys.micromanagertiff.micromanagertiffdatainterface import (
     MicroManagerTiffImagingInterface,
 )
@@ -151,12 +153,14 @@ interface_list = [
     # Ophys
     CaimanSegmentationInterface,
     CnmfeSegmentationInterface,
-    Suite2pSegmentationInterface,
     ExtractSegmentationInterface,
+    InscopixSegmentationInterface,
     SimaSegmentationInterface,
+    Suite2pSegmentationInterface,
     SbxImagingInterface,
     TiffImagingInterface,
     Hdf5ImagingInterface,
+    InscopixImagingInterface,
     ScanImageImagingInterface,
     ScanImageLegacyImagingInterface,
     ScanImageMultiFileImagingInterface,